Fix a few extra-pedantic warnings.

+++ source/plexlib/lexer/lexer.hpp
@@ -21,6 +21,6 @@ private:
 
 	std::string_view m_data;
 	std::queue<Token> m_buffer;
-	int m_lineNumber;
+	size_t m_lineNumber;
 	bool m_expectExpression;
 };
